+++ raiden-dapp/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [Unreleased]
+
+### Added
+- [#1365] Allow user to transfer ETH between Main and Raiden account.
+
 ### Fixed
 
 - [#1300] General screen shows Raiden account balance when using subkey.
@@ -8,6 +12,7 @@
 
 [#1300]: https://github.com/raiden-network/light-client/issues/1300
 [#1382]: https://github.com/raiden-network/light-client/issues/1382
+[#1365]: https://github.com/raiden-network/light-client/issues/1365
 
 ## [0.6.0] - 2020-04-21